Italian localization - Additions and updates
Added translations for these Extensions/Scripts:

Dynamic Prompts
Alpha Canvas
Artists to study
Aesthetic Score

Added a few missing translations and corrected some others. Updated to the latest Extension management tool version.

While I was able to translate in the statistics the text "Time taken:" because the timing itself is a separate label, I couldn't do the same with the labels "Torch active/reserved: 3881/3892 MiB" and "Sys VRAM: 4859/8192 MiB (59.31%)" because the values (memory, percentage) are embedded in the labels (or perhaps I am not enough acknowledged to be sure I am doing it right simply translating the text).